reactive user from front & new route/controller from back

+++ client/frontend/app/admin/page.jsx
@@ -4,12 +4,19 @@ import Cookies from 'js-cookie';
 import Link from 'next/link';
 import SearchUser from '@/components/SearchUser';
 import { useState, useEffect } from 'react';
-import { FaPlus, FaPen, FaRegTrashCan, FaUnlock } from 'react-icons/fa6';
+import {
+  FaPlus,
+  FaPen,
+  FaRegTrashCan,
+  FaLock,
+  FaUnlock,
+} from 'react-icons/fa6';
 import Loader from '@/components/Loader';
 
 export default function AdminPage() {
   const [users, setUsers] = useState([]);
   const [filteredUser, setFilteredUser] = useState([]);
+  const [userRoutines, setUserRoutines] = useState([]);
   const [userNameCookie, setUserNameCookie] = useState('');
   //const [loading, setLoading] = useState(true);
 
@@ -36,34 +43,78 @@ export default function AdminPage() {
     }
   };
 
+  const getAllRoutines = async () => {
+    try {
+      const res = await axios.get('http://localhost:8000/routines/all');
+      setUserRoutines(res.data);
+    } catch (error) {
+      console.error('Error fetching routines:', error);
+    }
+  };
+
   useEffect(() => {
     getAllUsers();
+    getAllRoutines();
     //Para no tener problema de hidratacion con el client component:
     setUserNameCookie(Cookies.get('userName'));
   }, []);
 
-  const searchOneUser = (searchString) => {
-    let userFounded = users.filter((el) =>
-      `${el.name} ${el.lastname}`
-        .toLowerCase()
-        .includes(searchString.toLowerCase())
+  const checkUserHasRoutine = (userId) => {
+    return userRoutines.some(
+      (routine) => routine.userId === userId && routine.userHasRoutine
     );
-    setFilteredUser(userFounded);
   };
 
-  // const handleAddRoutineById = async (id) => {
-  //   console.log(id);
+  const usersWithKeyHasRoutine = filteredUser.map((user) => ({
+    ...user,
+    hasRoutine: checkUserHasRoutine(user.id),
+  }));
 
-  //   try {
-  //     const { data } = await axios.get(`http://localhost:8000/users/${id}`);
-  //     console.log(data);
-  //   } catch (error) {
-  //     console.log(error);
-  //   }
-  // };
+  // setUsers(usersWithKeyHasRoutine);
+  // setFilteredUser(usersWithKeyHasRoutine);
 
-  const handleDisabledUser = async (id) => {
+  console.log('usuarios con la key hasRoutine:', usersWithKeyHasRoutine);
+
+  const searchOneUser = (searchString) => {
+    if (searchString.trim() === '') {
+      setFilteredUser(users);
+    } else {
+      let userFounded = usersWithKeyHasRoutine.filter((el) =>
+        `${el.name} ${el.lastname}`
+          .toLowerCase()
+          .includes(searchString.toLowerCase())
+      );
+      setFilteredUser(userFounded);
+    }
+  };
+
+  const handleReactiveUser = async (id, name, lastname) => {
     console.log(id);
+    const confirmReactive = window.confirm(
+      `Estas seguro de reactivar el usuario: ${name} ${lastname}?`
+    );
+    if (!confirmReactive) return;
+
+    try {
+      const res = await axios.put(
+        `http://localhost:8000/users/reactive/${id}`,
+        {
+          isActive: true,
+        }
+      );
+      console.log(res);
+      await getAllUsers();
+    } catch (error) {
+      console.error(`Error al activar usuario con el id ${id}:`, error);
+    }
+  };
+
+  const handleInactiveUser = async (id, name, lastname) => {
+    //console.log(id);
+    const confirmDelete = window.confirm(
+      `Estas seguro de inactivar el usuario: ${name} ${lastname}?`
+    );
+    if (!confirmDelete) return;
 
     try {
       const res = await axios.delete(`http://localhost:8000/users/${id}`, {
@@ -94,60 +145,83 @@ export default function AdminPage() {
       {/* Mapeo de usuarios */}
       <div className="grid grid-cols-2 gap-4">
         {users.length === 0 ? (
-          //TODO: cambiar el Loader por como lo hice en '/home'
           <div className="w-full flex justify-center items-center">
             <Loader />
           </div>
         ) : filteredUser.length === 0 ? (
           <p>No se encontro el usuario</p>
         ) : (
-          filteredUser.map((el) => (
-            <div
-              key={el.id}
-              className={`${
-                el.isActive === false
-                  ? 'flex flex-col justify-center items-center border border-primary  text-primary rounded-md p-2 gap-2 opacity-60'
-                  : 'flex flex-col justify-center items-center border border-primary text-primary rounded-md p-2 gap-2'
-              }`}
-            >
-              <p className="font-bold">
-                {el.name} {el.lastname}
-              </p>
-              <p className="text-center">
-                Creado el: {el.createdAt.split('T')}
-              </p>
-              {el.isActive === true ? (
-                <p className="font-bold">Estado: Con rutina</p>
-              ) : (
-                <p className="font-bold">Estado: Sin rutina</p>
-              )}
-              <div className="flex flex-row justify-center items-center gap-6">
+          usersWithKeyHasRoutine
+            .filter((el) => el.isAdmin !== true)
+            .map((el) => (
+              <div
+                key={el.id}
+                className={`${
+                  el.isActive === false
+                    ? 'flex flex-col justify-center items-center border border-primary  text-primary rounded-md p-2 gap-2 opacity-60'
+                    : 'flex flex-col justify-center items-center border border-primary text-primary rounded-md p-2 gap-2'
+                }`}
+              >
+                <p className="font-bold">
+                  {el.name} {el.lastname}
+                </p>
+                <p className="text-center">
+                  Se registro el: {el.createdAt.split('T')}
+                </p>
                 {el.isActive === true ? (
-                  <Link href={`/admin/create/${el.id}`} title="Editar rutina">
-                    <FaPen size={20} className="cursor-pointer" />
-                  </Link>
+                  <p className="font-bold">Estado: Activo</p>
                 ) : (
-                  <Link href={`/admin/create/${el.id}`} title="Crear rutina">
-                    <FaPlus size={20} className="cursor-pointer" />
-                  </Link>
+                  <p className="font-bold">Estado: Inactivo</p>
                 )}
-                {el.isActive === true ? (
-                  <FaRegTrashCan
-                    size={20}
-                    className="cursor-pointer"
-                    title="Inactivar usuario"
-                    onClick={() => handleDisabledUser(el.id)}
-                  />
-                ) : (
-                  <FaUnlock
-                    size={20}
-                    className="cursor-pointer"
-                    title="Activar usuario"
-                  />
-                )}
+                <div className="flex flex-row justify-center items-center gap-6">
+                  {el.hasRoutine === true ? (
+                    <Link href={`/admin/create/${el.id}`} title="Editar rutina">
+                      {el.isActive && (
+                        <FaPen size={20} className="cursor-pointer" />
+                      )}
+                    </Link>
+                  ) : (
+                    <>
+                      {el.isActive && (
+                        <Link
+                          href={`/admin/create/${el.id}`}
+                          title="Crear rutina"
+                        >
+                          <FaPlus size={20} className="cursor-pointer" />
+                        </Link>
+                      )}
+                    </>
+                  )}
+                  {el.isActive === true ? (
+                    <>
+                      <FaLock
+                        size={20}
+                        className="cursor-pointer"
+                        title="Inactivar usuario"
+                        onClick={() =>
+                          handleInactiveUser(el.id, el.name, el.lastname)
+                        }
+                      />
+                      <FaRegTrashCan
+                        size={20}
+                        className="cursor-pointer"
+                        title="Eliminar usuario"
+                        //onClick={() => handleDisabledUser(el.id)}
+                      />
+                    </>
+                  ) : (
+                    <FaUnlock
+                      size={20}
+                      className="cursor-pointer"
+                      title="Activar usuario"
+                      onClick={() =>
+                        handleReactiveUser(el.id, el.name, el.lastname)
+                      }
+                    />
+                  )}
+                </div>
               </div>
-            </div>
-          ))
+            ))
         )}
       </div>
     </div>
@@ -188,3 +262,14 @@ export default function AdminPage() {
 //   const data = await res.json();
 //   return data;
 // }
+
+// const handleAddRoutineById = async (id) => {
+//   console.log(id);
+
+//   try {
+//     const { data } = await axios.get(`http://localhost:8000/users/${id}`);
+//     console.log(data);
+//   } catch (error) {
+//     console.log(error);
+//   }
+// };
